refactor upload to S3
